Separate websocket message handler
